# SimBlog17API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SimBlog17API",
    "version": "1.0.0",
    "description": "Lightweight forum backend API.\nSupports JWT access/refresh tokens, articles, comments, interactions, homepage aggregation, and admin moderation.\n"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "paths": {
    "/api/user/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAccessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseUserAccessResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
    "/api/user/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAccessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseUserAccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
    "/api/user/auth/profile/{uid}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get user profile",
        "parameters": [
          {
            "$ref": "#/components/parameters/UidPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseUserAccessResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Update username",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UidPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsernameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseUserAccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
    "/api/user/auth/profile/{uid}/avatar": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Upload user avatar",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UidPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseMediaUploadResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/user/auth/password": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseBoolean"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
    "/api/user/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh tokens",
        "parameters": [
          {
            "$ref": "#/components/parameters/RefreshTokenHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseUserAccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/user/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout and revoke tokens",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AuthorizationHeader"
          },
          {
            "$ref": "#/components/parameters/RefreshTokenHeaderOptional"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseBoolean"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/user/auth/search/by_username": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Search users by username (plain text)",
        "parameters": [
          {
            "in": "query",
            "name": "keyword",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageUserAccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/auth/search/regex/by_username": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Search users by username (regex)",
        "parameters": [
          {
            "in": "query",
            "name": "pattern",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageUserAccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "InvalidRequest": {
        "description": "Invalid Request - validation or payload format error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusResponseString"
            },
            "example": {
              "statusCode": 2,
              "statusMessage": "Invalid Request",
              "timeStamp": 1778003000000,
              "traceId": "f84f2d4d-7f13-4bd1-8b63-ec2f58b1f54b",
              "data": "Validation failed: username must not be blank"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - missing/invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusBase"
            },
            "example": {
              "statusCode": 5,
              "statusMessage": "Unauthorized",
              "timeStamp": 1778002000000,
              "traceId": "5f8a6ca2-b8bf-4ad4-b3d4-0d7f7dd6a20b"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusBase"
            },
            "example": {
              "statusCode": 4,
              "statusMessage": "Access Denied",
              "timeStamp": 1778002000000,
              "traceId": "2dbf3d1d-06ee-4a8d-a865-1e8cc7f43a6d"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Too Many Requests - rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusBase"
            },
            "example": {
              "statusCode": 6,
              "statusMessage": "Too Many Requests",
              "timeStamp": 1778002000000,
              "traceId": "8a5fe15b-8f2b-4b7a-89df-4f44a13d8d7e"
            }
          }
        }
      }
    },
    "parameters": {
      "UidPath": {
        "in": "path",
        "name": "uid",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
      "Page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "Size": {
        "in": "query",
        "name": "size",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 200
        }
      },
      "Sort": {
        "in": "query",
        "name": "sort",
        "schema": {
          "type": "string",
          "example": "publishedTime,desc"
        }
      },
      "AuthorizationHeader": {
        "in": "header",
        "name": "Authorization",
        "required": false,
        "schema": {
          "type": "string",
          "example": "Bearer <access-token>"
        }
      },
      "RefreshTokenHeader": {
        "in": "header",
        "name": "X-Refresh-Token",
        "required": true,
        "schema": {
          "type": "string",
          "example": "<refresh-token>"
        }
      },
      "RefreshTokenHeaderOptional": {
        "in": "header",
        "name": "X-Refresh-Token",
        "required": false,
        "schema": {
          "type": "string",
          "example": "<refresh-token>"
        }
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ]
      },
...
      "UserAccessRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32
          },
          "password": {
            "type": "string",
            "maxLength": 128
          }
        }
      },
      "UpdateUsernameRequest": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 32
          }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "oldPassword",
          "newPassword"
        ],
        "properties": {
          "oldPassword": {
            "type": "string",
            "maxLength": 128
          },
          "newPassword": {
            "type": "string",
            "maxLength": 128
          }
        }
      },
      "UserAccessResponse": {
        "type": "object",
        "example": {
          "id": 1001,
          "username": "demo_user",
          "avatarUrl": "/api/media/files/avatar_1001.png",
          "role": "USER",
          "createTime": "2026-05-05 10:00:00",
          "accessToken": "<access-token>",
          "refreshToken": "<refresh-token>"
        },
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "MediaUploadResponse": {
        "type": "object",
        "example": {
          "id": 501,
          "url": "/api/media/files/20260505_abc123.png",
          "originalFileName": "cover.png",
          "storedFileName": "20260505_abc123.png",
          "contentType": "image/png",
          "sizeBytes": 102400
        },
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "storedFileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PageUserAccessResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAccessResponse"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
...
      "ApiStatusBase": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "statusMessage": {
            "type": "string"
          },
          "timeStamp": {
            "type": "integer",
            "format": "int64"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "ApiStatusResponseBoolean": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "ApiStatusResponseString": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponseUserAccessResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/UserAccessResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponseMediaUploadResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/MediaUploadResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponsePageUserAccessResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/PageUserAccessResponse"
              }
            }
          }
        ]
      },
...
    }
  }
}